Add: secure views with @login

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -75,8 +76,13 @@ def product_detail(request, artwork_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ Add artwork to the gallery """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, this is for authorised users only')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -97,8 +103,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, artwork_id):
     """ Edit artwork in the gallery """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, this is for authorised users only')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=artwork_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -122,8 +133,13 @@ def edit_product(request, artwork_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, artwork_id):
     """ Delete artwork from the gallery """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, this is for authorised users only')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=artwork_id)
     product.delete()
     messages.success(request, 'Artwork deleted')
